refactor: rename module + visibility
