+++ src/types/array/mod.rs
@@ -7,8 +7,8 @@ use super::IAllocator;
 use crate::VoidPtr;
 use crate::raw::root::RED4ext as red;
 
-mod r#static;
-pub use r#static::StaticArray;
+mod static_array;
+pub use static_array::StaticArray;
 
 /// A dynamically sized array.
 #[repr(transparent)]
